[SPOT-274][REFACTOR] Move Tests to Study Domain
